Merged PR 118493: Extend assets-summary route to return in use and not in use assets
# Justification
For the summary panel in the UI, we need to extend the assets-summary route so that we will have less requests for data to be populated in the panel.

# Implementation
I have added two new properties to the `AssetsCount` model to extend it to support `inUse` and `notInUse`. I have also changed the way we compute the values in GetAssetsCount, so that we won't need to many queries on the db, and work with the data in memory.

# Testing
I have added an integration test and also ran all the tests in the AssetSummaryTest class.

# Checklist
- [X] I tested changes to product code in product
- [ ] I considered updates to the wiki
- [ ] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product