Merged PR 120664: The query-assets take should not have a max value.
# Justification
For the previous releases setting no take to the query-assets route would get all the assets. Right now, the take defaults to 1000. The problem was that 1000 was also the maximum allowed when creating the request via the generated API. This was making it impossible for the clients to query more than 1000 assets. This means that the notebooks were only taking into consideration the first 1000 assets. (The 20.6 released version was doing query-assets with no take and was getting all assets, so this is not a released issue)

# Implementation
Removed the maximum value from the take specification in niapm.yml.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki
- [ ] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product

The query-assets take should not have a max value.