Merged PR 125573: Added support for specifying discovery type when creating assets
# Justification
As the service will handle creation of assets via the Create-Assets route, the request model for it needs to support specifying the discovery type

# Implementation
Added the discovery type enum to the request and mapper.

# Testing
Added an integration test.

# Checklist
- [X] I tested changes to product code in product
- [ ] I considered updates to the wiki
- [ ] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product

Added support for specifying discovery type when creating assets

Related work items: #1324165