Merged PR 122932: Support timestamp rules for current time comparisons
# Justification
#1224376
This closes out the remaining task in this story - adding support for current time timestamp tag rules.

# Implementation
####General strategy:
- Support timespans as a valid set point for timestamp tag conditions
- When evaluating the condition, compare the tag's value to the server's time +- the setpoint. (e.g. 02:00:00 means two hours past the current time)
- Since the passage of time can cause these rules to become active, we need to evaluate them on a regular interval

####Details:
In ConditionRecord.cs, setpoints are converted to a TypedValue that matches the tag's type. The assumption that the types will be the same doesn't hold for timespan setpoints, so I short-circuited the conversion in SetTypedValueFields. Then when evaluating timestamp rules, we can get the setpoint from the untyped list when it's a timespan. It's a little gross, but creating a new interface for typing setpoints would be a ton of overhead, and this is probably the only case we'll ever have where the set point type doesn't match the tag type.

When parsing timespans, I'm manually passing in a culture. Otherwise, the input format could change depending on where the server is installed.

The TagRuleEngine now uses the Timer service to schedule callbacks to evaluate relative time rules. The interval length is stored in the config file for the service. For this task, we want the callback to be a work subscriber, but the timer service currently only supports instance subscribers. Fixing this was a pretty simple change - I added a boolean to the RegisterCallback message that tells the timer service to route the response message back to the subscriber message's origin (which is just the service's name).

# Testing
- Built and tested the changes on a fresh 21.0 SystemLink install
- Added a new test to the ATS

# Checklist
- [x] I tested changes to product code in product
- [x] I considered updates to the wiki
- [x] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product

Related work items: #1224376