Merged PR 127478: Create rules for the last updated date for any tags
# Justification

#1224749

# Implementation
Added a new tag data type of `TAG_TIME`. For rules with this type, we use a tag's last updated timestamp as the value to compare against. Everything else works the same as with `DATE_TIME` tags. To keep the types in ConditionRecord the same, I manually instantiate a TypedValue and set its value to the tag's timestamp before continuing with the evaluation. This is a little gross because TypedValue only has an empty constructor and is probably not intended to be manually created, so I'm open to suggestions.
Normally, rules will only match against tag's with the same type as the rule, but `TAG_TIME` rules will match tags of any type. The way this all works is unfortunately opaque - the TagRuleEngine service queries for all tags that match a rule's search path without taking tag types into consideration. Instead, the `evaluate` method in ConditionRecord is called for every matching tag, but will return an error (which never gets bubbled up anywhere) if the rule's setpoints don't match the incoming value's type. For `TAG_TIME` rules, every tag update will always include a timestamp, so that check never fails.

This PR also includes a tiny bug fix for #1275650 to not evaluate disabled rules. There was already a Debug.Assert for this case, so it was easy to return a 'none' evaluation result instead.

# Testing
- Built & tested on a fresh install of 21.0 SystemLink
- Added a test to the ATS for last updated time rules

# Checklist
- [x] I tested changes to product code in product
- [x] I considered updates to the wiki
- [x] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product