Merged PR 143611: Added check for all system properties updated
# Justification
when a system was connected for the first time, the minion updated broadcast contains all 3 properties (connected, grains, sysapi) as properties which got updated, but the strategy for new clients did not have the case where all 3 were changed.

# Testing
Tested in the product.

# Checklist
- [X] I tested changes to product code in product
- [ ] I considered updates to the wiki
- [ ] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product

Added check for all system properties updated