Merged PR 145036: This is a proposed fix for the bug 1355793 allowing old python clients to sta...
# Justification
The OpenAPI description of contains an wrong
`operationId: getChannelPropertiesCount`
at the
  `/v1/files/{fileId}/groups:`
that should be named getGroups according to getChannels later on.
In the generated python client stubs the method is called getChannelPropertiesCount but returns the groups.

# Implementation
Duplicate the entry and make a redirect to make sure python scripts using the old operationId do not break.
Pick a path name that makes sure this entry is just there because of compatibility.
  `/v0/files/{fileId}/groups:`

# Related BUG

This is a proposed fix for the bug 1355793 allowing old python clients to stay stable with the generated stub.

# Decision

Because we do not believe that users are out there we just fix the OperationId

Related work items: #1355793