Merged PR 157149: Export notebooks as HTML or PDF
# Justification

Added support for exporting notebooks as HTML or PDF and excluding code input cells from the export.

#1411711
#1411744

# Implementation
HTML export is implemented via the `HTMLExporter` of `nbconvert` and works out of the box. PDF export is implemented via the `WebPDFExporter` of `nbconvert` for which we extended the Python package with pyppeteer and chromium in [this PR](https://dev.azure.com/ni/DevCentral/_git/Skyline/commit/0f17c51a7e580093009c81f037ab29d000e675ae?refName=refs%2Fheads%2Fmaster).

Because we updated the `nbconvert` dependency, I also had to update `papermill` and change the implementation of the `NINBExecEngine`. Because of this [breaking change](https://github.com/nteract/papermill/pull/472/files), the previous used `PapermillExecutePreprocessor` is no longer available. I re-implemented the logic to inject environment variables by using the new `PapermillNotebookClient`.

# Testing
Build the service locally and deployed it to a SystemLink Server installation. I verified that notebooks can be executed and the SystemLink specific environment variables are still injected. I also verified that notebooks can be exported as HTML and PDF, with or without excluding the code input cells.

I will add ATS tests for this in another PR.

# Checklist
- [x] I tested changes to product code in product
- [ ] I considered updates to the wiki
- [ ] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product

Related work items: #1411711, #1411744, #1446393