Merged PR 170365: Fix description in nifile.yml
# Justification
When adding the workspace supported recently, I had increased the operation version of listFiles from  3 to 4 initially. Later I decided to keep that version and add a queryFiles operation explicitly. Started of with the version number 2 to make a distinction between implementations with and without workspace support. But I forgot to change the description in the yml file.