Merged PR 172517: Increase delete-files operation version
# Justification
We're about to add this support to the LV API and without increasing the version number, we're not able to tell if force is supported by the backend. This enables us to check for support and print a meaningful error message.

'Force' has just recently been added and wasn't release yet.