Merged PR 194959: Make corrections to test monitor swagger documentation
# Justification

#1573106
#1576958

# Implementation
Based on comments from Ryan on the two bugs, I am updating the documentation to remove options that are listed as valid but are actually unsupported. The first bug has a corresponding code change that is tracked separately.

# Testing
Will verify that swagger docs list this change after submission.

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki
- [ ] If this PR affects user-visible strings, I added Jenna Jaco as a reviewer and attached screenshots of the strings in the product

Related work items: #1573106, #1576958