Merged PR 223882: Updated documentation in Swagger for query-assets route
# Justification

#1632322

# Implementation
We've enabled Dynamic LINQ filtering when requesting CSV reports for assets through the query-assets route. The changes have been made with this PR: !223432. Please make sure you read the description of that review for more context on the updated behavior. This has been agreed by the PO.

# Testing
N/A

# Checklist
- [ ] I tested changes to product code in product
- [ ] I considered updates to the wiki
- [ ] If this PR affects user-visible strings, I added the relevant tech writer as a reviewer and attached screenshots of the strings in the product

Updated documentation in Swagger for query-assets route

Related work items: #1632322